cms/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-braces==1.15.0
django-debug-toolbar==4.4.2
django-embed-video==1.4.10
djangorestframework==3.15.1
filelock==3.14.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
nodeenv==1.9.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg==3.1.19
pymemcache==4.0.0
pytest==8.2.1
pytest-django==4.8.0
python-decouple==3.8
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
virtualenv==20.26.2
